revert chaos webpage; use flatten in example
