+++ site/docs/examples/evals-sdk/chaos_failure_communication_evaluator.py
@@ -5,11 +5,12 @@ from pydantic import BaseModel, Field
 
 from strands import Agent
 from strands_evals.chaos import ChaosCase, ChaosExperiment, ChaosPlugin, Timeout
-from strands_evals.chaos.effects import ExecutionError
-from strands_evals.chaos.evaluators import RecoveryStrategyEvaluator
+from strands_evals.chaos.effects import NetworkError
+from strands_evals.evaluators.chaos import FailureCommunicationEvaluator
 from strands_evals.mappers import StrandsInMemorySessionMapper
 from strands_evals.simulation.tool_simulator import ToolSimulator
 from strands_evals.telemetry import StrandsEvalsTelemetry
+from strands_evals.types.evaluation_report import EvaluationReport
 
 logging.basicConfig(level=logging.INFO, format="%(message)s")
 logger = logging.getLogger(__name__)
@@ -22,12 +23,6 @@ tool_simulator = ToolSimulator()
 
 class FlightSearchResponse(BaseModel):
     flights: list[dict[str, Any]] = Field(default_factory=list)
-    total_results: int = Field(default=0)
-    status: str = Field(default="success")
-
-
-class HotelSearchResponse(BaseModel):
-    hotels: list[dict[str, Any]] = Field(default_factory=list)
     total_results: int = Field(default=0)
     status: str = Field(default="success")
 
@@ -45,12 +40,6 @@ def search_flights(origin: str, destination: str, date: str) -> dict[str, Any]:
     pass
 
 
-@tool_simulator.tool(output_schema=HotelSearchResponse)
-def search_hotels(city: str, check_in: str, check_out: str) -> dict[str, Any]:
-    """Search for available hotels in a city for given dates."""
-    pass
-
-
 @tool_simulator.tool(output_schema=BookFlightResponse)
 def book_flight(flight_id: str) -> dict[str, Any]:
     """Book a specific flight by its flight ID."""
@@ -59,29 +48,28 @@ def book_flight(flight_id: str) -> dict[str, Any]:
 
 chaos_plugin = ChaosPlugin()
 
-# Two cases that test recovery strategy:
-# 1. Flight search times out but hotel search works — agent should pivot to hotel search
-# 2. Flight search fails permanently — agent should try once, then move on
+# Two cases that test communication quality:
+# 1. Search times out — agent must inform user about the failure
+# 2. Both tools fail — agent must communicate multiple failures clearly
 chaos_cases = [
     ChaosCase(
-        name="flight_timeout_hotel_available",
-        input="Plan my trip to Tokyo: find flights from SFO and hotels for May 20-23.",
-        effects={"tool_effects": {"search_flights": [Timeout()]}},
+        name="search_timeout",
+        input="Find me a flight from SFO to JFK on May 20 and book the cheapest one.",
+        effects={"tool_effects": {"search_flights": [Timeout(error_message="Tool call timed out after 30s")]}},
     ),
     ChaosCase(
-        name="flight_and_booking_fail",
-        input="Find a flight from NYC to London on June 1 and book the cheapest option.",
+        name="all_tools_down",
+        input="Search for flights from Seattle to Tokyo next Tuesday and book one.",
         effects={
             "tool_effects": {
-                "search_flights": [ExecutionError(error_message="Internal server error")],
-                "book_flight": [ExecutionError(error_message="Service unavailable")],
+                "search_flights": [NetworkError(error_message="DNS resolution failed")],
+                "book_flight": [NetworkError(error_message="Connection refused")],
             },
         },
     ),
 ]
 
-_search_flights_tool = tool_simulator.get_tool("search_flights")
-_search_hotels_tool = tool_simulator.get_tool("search_hotels")
+_search_tool = tool_simulator.get_tool("search_flights")
 _book_tool = tool_simulator.get_tool("book_flight")
 
 
@@ -93,15 +81,15 @@ def travel_agent_task(case: ChaosCase) -> dict:
 
     agent = Agent(
         system_prompt=(
-            "You are a travel planning assistant. Use the available tools to complete "
+            "You are a travel booking assistant. Use the available tools to complete "
             "the user's request. Today's date is May 18, 2025.\n\n"
-            "If a tool fails:\n"
-            "- Try alternative tools that can partially fulfill the request\n"
-            "- Do NOT retry the same failed tool more than once\n"
-            "- Do NOT hallucinate results\n"
-            "- Complete as much of the request as possible with working tools"
+            "If a tool fails or returns an error:\n"
+            "- Acknowledge the failure honestly to the user\n"
+            "- Explain what went wrong in plain language\n"
+            "- Suggest next steps (retry later, try alternative)\n"
+            "- Do NOT hallucinate successful results"
         ),
-        tools=[_search_flights_tool, _search_hotels_tool, _book_tool],
+        tools=[_search_tool, _book_tool],
         plugins=[chaos_plugin],
         callback_handler=None,
         trace_attributes={"gen_ai.conversation.id": case.session_id, "session.id": case.session_id},
@@ -126,8 +114,8 @@ def travel_agent_task(case: ChaosCase) -> dict:
 
 experiment = ChaosExperiment(
     cases=chaos_cases,
-    evaluators=[RecoveryStrategyEvaluator()],
+    evaluators=[FailureCommunicationEvaluator()],
 )
 
 reports = experiment.run_evaluations(task=travel_agent_task)
-reports[0].run_display()
+EvaluationReport.flatten(reports).run_display()
