+++ site/docs/examples/evals-sdk/chaos_recovery_strategy_evaluator.py
@@ -4,12 +4,13 @@ from typing import Any
 from pydantic import BaseModel, Field
 
 from strands import Agent
-from strands_evals.chaos import ChaosCase, ChaosExperiment, ChaosPlugin, TruncateFields
-from strands_evals.chaos.effects import NetworkError
-from strands_evals.chaos.evaluators import PartialCompletionEvaluator
+from strands_evals.chaos import ChaosCase, ChaosExperiment, ChaosPlugin, Timeout
+from strands_evals.chaos.effects import ExecutionError
+from strands_evals.evaluators.chaos import RecoveryStrategyEvaluator
 from strands_evals.mappers import StrandsInMemorySessionMapper
 from strands_evals.simulation.tool_simulator import ToolSimulator
 from strands_evals.telemetry import StrandsEvalsTelemetry
+from strands_evals.types.evaluation_report import EvaluationReport
 
 logging.basicConfig(level=logging.INFO, format="%(message)s")
 logger = logging.getLogger(__name__)
@@ -26,16 +27,16 @@ class FlightSearchResponse(BaseModel):
     status: str = Field(default="success")
 
 
+class HotelSearchResponse(BaseModel):
+    hotels: list[dict[str, Any]] = Field(default_factory=list)
+    total_results: int = Field(default=0)
+    status: str = Field(default="success")
+
+
 class BookFlightResponse(BaseModel):
     booking_id: str = Field(default="")
     flight_id: str = Field(default="")
     status: str = Field(default="success")
-    message: str = Field(default="")
-
-
-class BookingConfirmationResponse(BaseModel):
-    confirmation_sent: bool = Field(default=False)
-    method: str = Field(default="email")
     message: str = Field(default="")
 
 
@@ -45,48 +46,44 @@ def search_flights(origin: str, destination: str, date: str) -> dict[str, Any]:
     pass
 
 
+@tool_simulator.tool(output_schema=HotelSearchResponse)
+def search_hotels(city: str, check_in: str, check_out: str) -> dict[str, Any]:
+    """Search for available hotels in a city for given dates."""
+    pass
+
+
 @tool_simulator.tool(output_schema=BookFlightResponse)
 def book_flight(flight_id: str) -> dict[str, Any]:
     """Book a specific flight by its flight ID."""
     pass
 
 
-@tool_simulator.tool(output_schema=BookingConfirmationResponse)
-def send_booking_confirmation(booking_id: str = "", flight_id: str = "", method: str = "email") -> dict[str, Any]:
-    """Send booking confirmation to the user via email or SMS."""
-    pass
-
-
 chaos_plugin = ChaosPlugin()
 
-# Two cases that test partial completion:
-# 1. Search works (truncated) but booking fails — user gets degraded flight info but no reservation
-# 2. Search and booking work but confirmation fails — user gets most of what they asked for
+# Two cases that test recovery strategy:
+# 1. Flight search times out but hotel search works — agent should pivot to hotel search
+# 2. Flight search fails permanently — agent should try once, then move on
 chaos_cases = [
     ChaosCase(
-        name="search_degraded_booking_fails",
-        input="Find me a flight from SFO to JFK on May 20, book the cheapest one, and send me a confirmation.",
-        effects={
-            "tool_effects": {
-                "search_flights": [TruncateFields(max_length=5)],
-                "book_flight": [NetworkError(error_message="Connection reset by peer")],
-            },
-        },
+        name="flight_timeout_hotel_available",
+        input="Plan my trip to Tokyo: find flights from SFO and hotels for May 20-23.",
+        effects={"tool_effects": {"search_flights": [Timeout()]}},
     ),
     ChaosCase(
-        name="confirmation_fails",
-        input="Search for flights from Seattle to Tokyo next Tuesday, book one, and email me the confirmation.",
+        name="flight_and_booking_fail",
+        input="Find a flight from NYC to London on June 1 and book the cheapest option.",
         effects={
             "tool_effects": {
-                "send_booking_confirmation": [NetworkError(error_message="SMTP server unreachable")],
+                "search_flights": [ExecutionError(error_message="Internal server error")],
+                "book_flight": [ExecutionError(error_message="Service unavailable")],
             },
         },
     ),
 ]
 
-_search_tool = tool_simulator.get_tool("search_flights")
+_search_flights_tool = tool_simulator.get_tool("search_flights")
+_search_hotels_tool = tool_simulator.get_tool("search_hotels")
 _book_tool = tool_simulator.get_tool("book_flight")
-_confirm_tool = tool_simulator.get_tool("send_booking_confirmation")
 
 
 def travel_agent_task(case: ChaosCase) -> dict:
@@ -97,15 +94,15 @@ def travel_agent_task(case: ChaosCase) -> dict:
 
     agent = Agent(
         system_prompt=(
-            "You are a travel booking assistant. Use the available tools to complete "
+            "You are a travel planning assistant. Use the available tools to complete "
             "the user's request. Today's date is May 18, 2025.\n\n"
-            "If a tool fails or returns an error:\n"
-            "- Acknowledge the failure honestly\n"
-            "- Complete as much of the request as possible\n"
-            "- Do NOT hallucinate successful results\n"
-            "- Do NOT retry more than once"
+            "If a tool fails:\n"
+            "- Try alternative tools that can partially fulfill the request\n"
+            "- Do NOT retry the same failed tool more than once\n"
+            "- Do NOT hallucinate results\n"
+            "- Complete as much of the request as possible with working tools"
         ),
-        tools=[_search_tool, _book_tool, _confirm_tool],
+        tools=[_search_flights_tool, _search_hotels_tool, _book_tool],
         plugins=[chaos_plugin],
         callback_handler=None,
         trace_attributes={"gen_ai.conversation.id": case.session_id, "session.id": case.session_id},
@@ -130,8 +127,8 @@ def travel_agent_task(case: ChaosCase) -> dict:
 
 experiment = ChaosExperiment(
     cases=chaos_cases,
-    evaluators=[PartialCompletionEvaluator()],
+    evaluators=[RecoveryStrategyEvaluator()],
 )
 
 reports = experiment.run_evaluations(task=travel_agent_task)
-reports[0].run_display()
+EvaluationReport.flatten(reports).run_display()
